Implemented get_all_files function
This function returns an array that includes paths to all files in the repository. Presently it does not exclude anythin (including .git and the januus metadata itself) this will obviously cause a lot of problems, but the intention of this project is to create the most barebone version control system possible, even though it will be non-polinomialy inefficient, it will save all the data available and as such it will be able to do anythin git does in the future (specifically talking about this repository), unfortunetaly there is a limited number of 'commits' that can be done using this MMMVP because I think there is a maxiumum path length in UNIX systems, but we will see considering I am also using .git I can always backtrack if it becomes to difficult (this is also a reason why I wont .gitignore .jan, because unfortunately then it won't be as 'fun').