Implementd copy_all_files function
This function copies all files from the repository (excluding files that are generated during its execution).
It is intentional that I do not copy the files that are generated during the execution of the function because that would cause an infinite recursion.
I have also cleared .gitignore to not create weird missing files in .jan.